Remove "integer to pointer without cast" warning where I set bpload_bp to the IRET address of the EXEC call, as read off the stack.
git-svn-id: https://dosemu.svn.sourceforge.net/svnroot/dosemu/branches/debugger@1887 7758b857-130e-0410-ab9c-d8735a528d18

+++ src/arch/linux/debugger/mhpdbg.c
@@ -349,8 +349,8 @@ unsigned int mhp_debug(enum dosdebug_event code, unsigned int parm1, unsigned in
 	    if ((mhpdbgc.bpload==1) && (DBG_ARG(mhpdbgc.currcode) == 0x21) && (LWORD(eax) == 0x4b00) ) {
 
 	      /* mhpdbgc.bpload_bp=((long)LWORD(cs) << 4) +LWORD(eip); */
-	      mhpdbgc.bpload_bp = MK_FP32(READ_WORD(SEG_ADR((Bit16u *), ss, sp) + 1),
-					  READ_WORD(SEG_ADR((Bit16u *), ss, sp) + 0));
+	      mhpdbgc.bpload_bp = (long) MK_FP32(READ_WORD(SEG_ADR((Bit16u *), ss, sp) + 1),
+						 READ_WORD(SEG_ADR((Bit16u *), ss, sp) + 0));
 
 	      if (mhp_setbp(mhpdbgc.bpload_bp)) {
 		mhp_printf("\n\nbpload: intercepting EXEC:\n", LWORD(cs), REG(eip));
